chore: rename sagas to calculationSaga
