+++ src/state/createStore.js
@@ -1,6 +1,6 @@
 import { applyMiddleware, combineReducers, compose, createStore } from 'redux';
 import createSagaMiddleware from 'redux-saga';
-import gearOptimizerSaga from './sagas/sagas';
+import calculationSaga from './sagas/calculationSaga';
 import formStateSaga from './sagas/formStateSaga';
 import buildPageSaga from './sagas/buildPageSaga';
 import { bossSlice } from './slices/boss';
@@ -44,7 +44,7 @@ const composeEnhancers =
 
 export default () => {
   const store = createStore(reducers, composeEnhancers(applyMiddleware(saga)));
-  saga.run(gearOptimizerSaga);
+  saga.run(calculationSaga);
   saga.run(formStateSaga);
   saga.run(buildPageSaga);
   return store;
